+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0-prerelease.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.11...@patternfly/react-docs@7.0.0-prerelease.12) (2024-09-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.10...@patternfly/react-docs@7.0.0-prerelease.11) (2024-09-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
 # [7.0.0-prerelease.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.9...@patternfly/react-docs@7.0.0-prerelease.10) (2024-09-11)
 
 **Note:** Version bump only for package @patternfly/react-docs